+++ packages/api-graphql/src/GraphQLAPI.ts
@@ -21,6 +21,7 @@ import Amplify, {
 	Credentials,
 	// @ts-ignore
 	INTERNAL_AWS_APPSYNC_PUBSUB_PROVIDER,
+	Constants,
 } from '@aws-amplify/core';
 import PubSub from '@aws-amplify/pubsub';
 import Auth from '@aws-amplify/auth';
@@ -28,6 +29,7 @@ import Cache from '@aws-amplify/cache';
 import { GraphQLOptions, GraphQLResult } from './types';
 import { v4 as uuid } from 'uuid';
 import { RestClient } from '@aws-amplify/api-rest';
+const USER_AGENT_HEADER = 'x-amz-user-agent';
 
 const logger = new Logger('GraphQLAPI');
 
@@ -221,8 +223,11 @@ export class GraphQLAPIClass {
 				(customEndpointRegion
 					? await this._headerBasedAuth(authMode)
 					: { Authorization: null })),
-			...additionalHeaders,
 			...(await graphql_headers({ query, variables })),
+			...additionalHeaders,
+			...(!customGraphqlEndpoint && {
+				[USER_AGENT_HEADER]: Constants.userAgent,
+			}),
 		};
 
 		const body = {