+++ packages/aws-amplify-react/src/browser.d.ts
@@ -1,10 +1,6 @@
 declare interface Document {
-	exitFullscreen: () => void;
 	mozCancelFullScreen: () => void;
 	webkitExitFullscreen: () => void;
-	fullscreenElement: () => void;
-	mozFullScreenElement: () => void;
-	webkitFullscreenElement: () => void;
 }
 
 declare interface Window {